# scaffold/contracts/openapi/m6-gateway.v0.3.yaml
# ============================================================================
# M6 统一数据出口（骨架版）契约摘要 v0.3
# ----------------------------------------------------------------------------
# 真源：服务运行后自动生成的 /openapi.json。本文件是它的人工摘要，用于评审与对接，
#       不参与运行时校验——避免"文档与实现两套真源"。
# 接入纪律：任何模块读数据必须经本服务；禁止跨模块直连数据库（见 MODULE-ONBOARDING.md）。
# ============================================================================
openapi: 3.1.0
info:
  title: M6 统一数据出口（骨架版）
  version: 0.3.0
  description: 对象查询 / 几何查询 / 指标查询 / 动作提交四类接口。
servers:
  - url: http://localhost:8001        # 骨架栈映射端口（容器内 8000）

paths:
  /healthz:
    get:
      summary: 健康检查
      responses:
        "200":
          description: 依赖状态

  /v1/catalog:
    get:
      summary: 对象域目录（M3 契约③ 在线视图）
      description: >-
        返回 7 大对象域（GE/SU/RE/LO/WE/TE/DE）及其已建物理表、仅逻辑视图存在的实体、
        落库位置，以及可用的对象类型白名单。前端与课题组据它发现有哪些对象可取，
        不必对照文档硬编码。真源为 M3 的 rpdao.catalog。
      responses:
        "200":
          description: domains / cross_tables / object_types

  /v1/catalog/tables:
    get:
      summary: 逐张逻辑表盘点：行数 / 分区根表 / 是否已有数据
      description: >-
        平台「到底有哪些表、哪些有数据」的唯一数据源。**数的是逻辑表，不是物理表**：
        wim_axle_record 是声明式分区表，物理上是 1 个父表 ＋ 若干月分区 ＋ 1 个兜底分区；
        分区是 PostgreSQL 的实现细节，不是平台对象类型，故只报父表一行，
        分区数与分区键单列在 partition_count / partition_key 字段里。
        判「有没有数据」用 row_count > 0。真源为 M3 的 catalog.ALL_TABLES 与
        repo.PARTITIONED_ROOTS；行数由 M3 现查（UNION ALL 一次查完，非逐表往返）。
      responses:
        "200":
          description: >-
            logical_table_count / with_data_count / empty_count /
            physical_partition_count / items[]（table, domain, row_count,
            has_data, is_partitioned, partition_key, partition_count）
        "503":
          description: 存储不可用

  /v1/catalog/gaps:
    get:
      summary: 逐张逻辑表的缺口归因：为什么空 / 归谁 / 先动哪条
      description: >-
        从「哪些表是空的」推进到「每条空链该谁接」。归因需要三类事实，
        分属两个模块，**谁也不能替谁**：
        · 表里有几行 / 表的归属模块 → M3 rpdao（平台内唯一接触存储处）；
        · 段实现了没 / 源收到没 → M2（解析器与导入台账都在它那边）。
        故本接口在 M6 做**纯拼装**：M3 出表事实、M2 出段事实，按「段 ↔ 表」合并。
        让 M3 去读磁盘、或让 M2 去查库都能更快写完，但都会破坏「平台内唯一接触
        存储处」这条硬线 —— 而那条线正是四组学生能并行开发的前提。
        
        ★ 最要紧也最容易判错的一处：**「源在不在」不是看磁盘，看 design_file 台账**。
        M2 的导入入口是上传，文件落在临时目录、请求结束即销毁，服务端没有持久的
        设计目录 —— 扫磁盘会**恒得 false**，把「源已收到」错判成「源缺失」，
        而且不报错。故 source_state 取自 M2 的 design_file（经 /v1/design/capabilities）。
        
        ★ 措辞边界：source_present 只表示**收到过该文件**，
        **不**表示「导入一定能成功」。对外文案只能说「源已收到/已具备」，
        **不得**说「可以导入」—— 后者是给用户一个会落空的承诺。
        
        本接口只读，不返回任何行内容，也不提供写操作。
      parameters:
        - name: empty_only
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: true = 只返回空表（有数据的表仍计入汇总计数，只是不在 items 里）
      responses:
        "200":
          description: >-
            logical_table_count / with_data_count / empty_count / unknown_count /
            reason_counts / items[]（table, domain, row_count, has_data,
            is_partitioned, partition_count, owner, phase, empty_reason,
            segments, suffixes, source_present, action, priority_keys）
            
            empty_reason 为封闭枚举：has_data / source_ready_not_imported /
            source_needs_conversion / source_absent / module_not_built /
            upstream_pending / unknown。扩枚举须走契约变更。
            
            priority_keys 是三键数组 [改动模块数, 源是否具备, 域序]，
            **必须**暴露：复核者要靠它独立复算排序（否则排序不可证伪）。
            同输入必得同一次序 —— 三键之后以表名字典序决胜。
            
            source_present 是三态：true / false / null（null = 该表不对应任何段，
            即下游产出表，与「源缺失」是两件事，不得压缩成 false）。
        "503":
          description: >-
            上游不可用，**并点名是哪一个**：M3 挂了是「所有数据都拿不到」，
            M2 挂了是「只有归因算不全」，两者排查方向完全不同。

  /v1/objects/{object_type}:
    get:
      summary: 按对象类型查档案对象
      parameters:
        - name: object_type
          in: path
          required: true
          schema:
            type: string
          description: >-
            白名单＝7 大对象域已建的物理表（共 21 张，完整清单见 GET /v1/catalog
            的 object_types）。跨域支撑表（字典 5 张 + 治理/闭环 4 张）刻意不经此暴露。
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 1000
      responses:
        "200":
          description: 返回 object_type、domain、count 与 items 数组
        "404":
          description: 未登记的对象类型

  /v1/objects/wim_axle:
    get:
      summary: WIM 过车记录（按时间 / 桩号 / 超载筛选）
      parameters:
        - name: from_ts
          in: query
          schema:
            type: string
            format: date-time
        - name: to_ts
          in: query
          schema:
            type: string
            format: date-time
        - name: station
          in: query
          schema:
            type: string
          example: K4640+000
        - name: overload_only
          in: query
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
      responses:
        "200":
          description: 返回 object_type、count 与 items 数组

  /v1/objects/wim_axle/{record_id}:
    get:
      summary: 单条过车记录 + 轴明细（对象→链接展开，为本体"对象-链接"图预留形态）
      parameters:
        - name: record_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 记录字段 + axles 明细数组
        "404":
          description: 记录不存在

  # ── 几何查询（工单 #4）────────────────────────────────────────────────────
  # GE 的数据不是"一堆并列对象"，而是**以 road_section 为根的一棵树**：桩号序列、
  # 平面交点、线形单元全部锚在路段上。平铺查「所有交点」在只有一个路段时看着没问题，
  # 路段一多就静默串台（两条路的桩号都从 0 开始，混了看不出来）。
  # 所以这里一律**按路段取子树**，不提供无 section 的取法。
  # 实现全部落在 M3 的 GeRepository（契约③），本服务只做翻译与筛选参数转发。
  /v1/geometry/sections:
    get:
      summary: 路段列表（含路线/设计项目/分段属性 + 三类几何计数）
      description: >-
        GE 域的入口。返回全部路段；每个路段带 station_count / pi_count / element_count，
        供前端在打开某个路段前就知道它有没有几何数据。
      responses:
        "200":
          description: count 与 items 数组

  /v1/geometry/sections/{section_id}/stations:
    get:
      summary: 桩号序列（可按区间与是否整桩筛）
      description: >-
        桩号是 GE 域的一等实体，其余逐桩数据全部锚在它上面，故这是几何浏览的主视图。
        返回 station_local_km（路段内里程）与 station_text（原样桩号文本）。
      parameters:
        - name: section_id
          in: path
          required: true
          schema:
            type: integer
        - name: from_km
          in: query
          schema:
            type: number
          example: 0.6
        - name: to_km
          in: query
          schema:
            type: number
        - name: integer_only
          in: query
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          schema:
            type: integer
            default: 500
      responses:
        "200":
          description: section_id、count 与 items 数组

  /v1/geometry/sections/{section_id}/alignment:
    get:
      summary: 平面线形（交点链 + 线形单元链，一次取回）
      description: >-
        pis 为交点（R / Ls / A / T / 转角 / 外距），elements 为线形单元
        （circular / transition / line）。element 的 pi_id 为空表示**直线段** ——
        直线不属于任何交点，这是正常的，不是漏挂。
      parameters:
        - name: section_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: section_id、pis 数组与 elements 数组

  /v1/geometry/sections/{section_id}/summary:
    get:
      summary: 单条路段的概要 + 几何完整度报告
      description: >-
        section 为该路段自身（含所属路线、设计项目与源文件名）；
        completeness 回答「为什么这个路段的几何等级是 L2 而不是 L3」——
        等级是由"哪些段有数据"当场推出的派生量，故连依据（每张表的行数）一并返回，
        让结论可被核对。缺口分 missing_not_built（schema 没到，如 v0.4 的横断面）
        与 missing_no_data（表建好了但解析器没做）两类，含义不同。
        注意：设计数据的**导入批次台账**（data_import_batch）没有指回路段的列，
        故这里给不出"哪个批次导的"，只能给到文件级来源。
      parameters:
        - name: section_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: section 对象与 completeness 对象
        "404":
          description: 路段不存在

  /v1/geometry/sections/{section_id}/widths:
    get:
      summary: 路幅宽度表（`.WID` 的变化点，按桩号）
      description: >-
        ⚠ 这是**变化点**表，不是逐桩表：一行 = 一个宽度分组的起点。
        实测毕设路段**只有 4 行**，而 `.WID` 覆盖 0.000–5701.461 m。
        调用方要拿"某个桩号的宽度"，须**按桩号取最后一行**
        （`<= 该桩号` 的 station_km 最大者）—— 那是业务语义，网关不替调用方决定。
        ⚠ 终点 5701.461 m **早于**路线终点 5805.421 m：后 103.960 m
        **源文件就没有**路幅宽度数据，这是源文件缺口，不是解析漏了。
        ⚠ side 取 left / right（全库 8 张 side 表统一词表）；源文件写的是
        [LEFT]/[RIGHT]，由解析器归一化。同一桩号左右各一行，group_seq 区分同侧多组。
      parameters:
        - name: section_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: section_id、count 与 items 数组（每行一个变化点）

  /v1/geometry/sections/{section_id}/superelevation:
    get:
      summary: 超高过渡（`.SUP` 的逐桩横坡控制点）
      description: >-
        ⚠ 这是**控制点**表，不是逐桩表（实测毕设 76 行）。
        **控制点之间横坡是线性渐变的** —— 想拿某个桩号的横坡必须插值，
        直接取"最后一行"是错的。这一条有实测支撑：`.lj` 的 11 个高差列
        能由本表复现，而**只有插值才对得上**（按"沿用上值"算有 55/332 行不符）。
        ⚠ 源文件的 **9999 表示"忽略此数据"**，落库为 **null**。
        语义是「横坡渐变**穿过**这个控制点继续走」，所以插值时要**跳过 null
        去找它两侧最近的非 null**，而不是当 0、也不是沿用前一个值。
        ⚠ **左右两侧用的是同一套符号约定**：正常路拱两侧都写 −2.00，
        超高段写 +4.00 / −4.00。所以"左比右高"不能直接比大小 ——
        要按 −σ·宽度·横坡 折算成高差（σ = +1 左 / −1 右）。
      parameters:
        - name: section_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: section_id、count 与 items 数组（每行一个控制点）
        "404":
          description: 路段不存在

  /v1/metrics/wim_hourly:
    get:
      summary: 小时级过车量 / 超载数 / ESAL / 均速
      parameters:
        - name: day
          in: query
          schema:
            type: string
            format: date
        - name: station
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 返回 passages、overloaded、overload_ratio、esal_sum 与 buckets 数组

  /v1/actions/{action_name}:
    post:
      summary: 受治理动作提交（骨架期返回 501，P3 落地）
      description: >-
        计划动作（报告 10.4 ③）：create_maintenance_ticket / dispatch_alert /
        confirm_event / mark_data_quality_issue / request_design_change。
        每个动作必须实现：参数校验 → 权限校验 → 副作用（通知 / 工单）→ 写回源系统 → 审计留痕，
        且默认在人工确认后才执行。
      parameters:
        - name: action_name
          in: path
          required: true
          schema:
            type: string
      responses:
        "501":
          description: 尚未实现（骨架期约定）

components:
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string